VNLP-1357 [Livechat][Maintenance] Gửi tin cho user thành công khi đang trong quá trình bảo trì

+++ socket-handler/message.js
@@ -21,6 +21,9 @@ exports.initEvent = (socket) => {
                     const engineId = socket.engine._id;
                     const orgId = socket.org._id;
 
+                    const isMaintenance = await messageService.verifyMaintenance(engineId);
+                    if (isMaintenance) return callback(new Error('System is under maintenance'));
+
                     await messageService.removeTimer(roomId, '*', engineId);
                     const { message, room } = await messageService.sendAgentMessage({
                         roomId,
